+++ CHANGELOG.md
@@ -14,3 +14,4 @@ Mechanisms, extensions and resources related to Symfony Flash Messages (https://
 8. Implement Phing
 9. Add PHPUnit configuration
 10. Add main class of bundle
+11. Add configuration of bundle